+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Other changes:
 - Added support for object de/serialization (`O:4:"name":n:{...}`).
 - Added `[PhpClass()]` attribute.
 - Added `StdClass` and `EnableTypeLookup` to deserialization options
+- Added options for `PhpSerialization.Serialize()`.
+	- `ThrowOnCircularReferences` - whether or not to throw on ciruclar references, defaults to false (this might change in the future.)
 - Updated and adjusted some of the XML documentation.
 
 # 0.4.0